profile page finish

+++ src/pages/profilePage/index.js
@@ -7,13 +7,14 @@ import {getProfile, updateProfile,deleteProfile} from '../../api/Profile';
 /* components */
 import { RiLoader3Line } from 'react-icons/ri';
 import {BiTrashAlt} from 'react-icons/bi';
-import Flash from '../../components/Flash';
 import ToolBox from '../../components/toolbox';
 import Loader from '../../components/Loader';
 import Header from '../../components/heading';
 import Form from '../../components/form';
 import Avatar from '../../components/avatar';
 import ImageLoader from '../../components/imageLoader';
+import Popup from '../../components/popup';
+import Flash from '../../components/flashMessage';
 
 //css
 import './ProfilePage.css';
@@ -38,11 +39,19 @@ const ProfilePage = (props) => {
         success: {},
         error: {}
     });
-    const [logout, setLogout] = useState(false);
+    const [flash, setFlash] = useState({
+        success: '',
+        error: ''
+    })
+    const [confirmPopup, setConfirmPopup] = useState({
+        isOpen: false,
+        action: null
+    })
 
     //refs
     const inputRef = useRef();
     const avatarRef = useRef();
+    const linkRef = useRef();
 
     //handlers
     const handleInputValue = debounce(function(text){
@@ -103,6 +112,24 @@ const ProfilePage = (props) => {
         }
     }
 
+    const handlePopupClick = function(e){
+        if(e.target.id === "confirm"){
+            return setConfirmPopup({
+                ...confirmPopup,
+                isOpen: false,
+                action: e.target.id
+            })
+        }
+
+        if(e.target.id === "cancel"){
+            return setConfirmPopup({
+                ...confirmPopup,
+                isOpen: false,
+                action: null
+            })
+        }
+    }
+
     const handleClick = debounce(function(e){
        switch(e.target.id){
            case "update":
@@ -112,16 +139,23 @@ const ProfilePage = (props) => {
                     type: "update"
                 });
             case "delete":
+                setConfirmPopup({
+                    ...confirmPopup,
+                    isOpen: true
+                })
                 return setUser({
                     ...user,
-                    pending:true,
-                    type: "delete",
+                    pending: true,
+                    type: "delete"
                 });
             default: 
                 return setUser({...user});
        }
     }, 400)
 
+    const logout = function(){
+        linkRef.current.click();
+    }
 
     useEffect(async () => {
         let response = '';
@@ -130,6 +164,11 @@ const ProfilePage = (props) => {
                 case "update":
                     response = await updateProfile({username: user.success.username, avatar: user.file});
                     if(response.code === 200){
+                        setFlash({
+                            ...flash,
+                            success: "Updated successfully",
+                            error: ""
+                        })
                         return setUser({
                             ...user,
                             pending: false,
@@ -137,18 +176,41 @@ const ProfilePage = (props) => {
                         })
                     }
                 case "delete":
-                    response = await deleteProfile();
-                    if(response.code === 200){
-                        setUser({
-                            ...user,
-                            pending:false,
-                            success:{...response.message}
-                        });
-                        return setLogout();
-                    }
+                    setUser({
+                        ...user,
+                        pending: false
+                    })
+                    return setConfirmPopup({
+                        ...confirmPopup,
+                        isOpen: true
+                    })
             }
         }
     },[user.pending])
+
+    useEffect(() => {
+        console.log(user)
+    },[user])
+
+    useEffect(async () => {
+        let response = '';
+        if(!confirmPopup.isOpen && confirmPopup.action === "confirm"){
+            response = await deleteProfile();
+            if(response.code === 200){
+                return setFlash({
+                     ...flash,
+                     error: '',
+                     success: response.message
+                })
+            }else{
+                return setFlash({
+                     ...flash,
+                     success:'',
+                     error: response.message
+                }) 
+            }
+        }
+    },[confirmPopup])
 
     useEffect(async () => {
         let responseUser = await getProfile();
@@ -175,6 +237,13 @@ const ProfilePage = (props) => {
                     <Header>
                         <Header.Logo url={kingdomsDatas[0].icon_url} alt="Logo 7fallen"/>
                     </Header>
+                    <Flash 
+                        error={flash.error} 
+                        success={flash.success} 
+                        setFlash={setFlash}
+                        redirect={user.type === "delete"}
+                        redirectCallback={logout} 
+                    />
                     <div className="avatar__block">
                         <Form.Label htmlFor="avatar__upload">
                             {
@@ -186,11 +255,12 @@ const ProfilePage = (props) => {
                             }
                             {
                                 user.pending === true &&
+                                user.type === "update" &&
                                     <Loader loaderIcon={RiLoader3Line}/>
                             }
                             {
-                                user.type === "update" &&
-                                user.pending === false &&  
+                                ((user.type === "delete" && user.pending === false) ||
+                                (user.type === "update" && user.pending === false)) &&  
                                     <Avatar 
                                         url={user.success.avatar && "/avatars" + user.success.avatar} 
                                         alt="user image"
@@ -219,6 +289,17 @@ const ProfilePage = (props) => {
                             debounceTime = {150}
                         />
                     </Form>
+                    <Popup isOpen={confirmPopup.isOpen} onClick={handlePopupClick}>
+                        <Popup.Text 
+                            text="Etes-vous certain de vouloir supprimer votre compte ? 
+                                  Cette action est irreversible et tous vos decks seront 
+                                  supprimés de la base de données."
+                        />
+                        <Popup.Group>
+                            <Popup.Button id="confirm" text="confirmer"/>
+                            <Popup.Button id="cancel" text="annuler"/>
+                        </Popup.Group>
+                    </Popup>
                     <ToolBox onClick={handleClick}>
                         <ToolBox.Content>
                             <ToolBox.Row>
@@ -227,10 +308,14 @@ const ProfilePage = (props) => {
                                     isLoading={user.pending && user.type === "update"} 
                                     text="mettre a jour"
                                 />
-                                <ToolBox.Action id="delete" icon={BiTrashAlt}/>
+                                <ToolBox.Action 
+                                    id="delete" 
+                                    icon={BiTrashAlt}
+                                />
                             </ToolBox.Row>
                         </ToolBox.Content>
                     </ToolBox>
+                    <a ref={linkRef} className="d-none" href={process.env.REACT_APP_BASE_URL + "/api/auth/logout"}></a>
                 </div>
                 :
                 <Loader loaderIcon={RiLoader3Line}/>
